Added basic React app structure and basic location filter
There's now an App, Sidebar, Matrix structure, to build upon
which allows the Sidebar to do the filtering, and then pass the filtered data
to the matrix.
Added a basic location filter, for testing.
Improved structure of Matrix.jsx and MatrixScale.jsx, added a file
MatrixUtility.js with static arrays/functions to reduce clutter a bit.